Update PT build scripts to work on Intel Macs too

+++ desktop/scripts/build-pt-meek.sh
@@ -10,10 +10,14 @@ cd meek
 git checkout $MEEK_TAG
 
 if [ "$OS" == "Darwin" ]; then
-    go build -o ../../../onionshare/resources/tor/meek-client-arm64 ./meek-client
-    GOOS=darwin GOARCH=amd64 go build -o ../../../onionshare/resources/tor/meek-client-amd64 ./meek-client
-    lipo -create -output ../../../onionshare/resources/tor/meek-client ../../../onionshare/resources/tor/meek-client-arm64 ../../../onionshare/resources/tor/meek-client-amd64
-    rm ../../../onionshare/resources/tor/meek-client-arm64 ../../../onionshare/resources/tor/meek-client-amd64
+    if [[ $(uname -m) == 'arm64' ]]; then
+        go build -o ../../../onionshare/resources/tor/meek-client-arm64 ./meek-client
+        GOOS=darwin GOARCH=amd64 go build -o ../../../onionshare/resources/tor/meek-client-amd64 ./meek-client
+        lipo -create -output ../../../onionshare/resources/tor/meek-client ../../../onionshare/resources/tor/meek-client-arm64 ../../../onionshare/resources/tor/meek-client-amd64
+        rm ../../../onionshare/resources/tor/meek-client-arm64 ../../../onionshare/resources/tor/meek-client-amd64
+    elif [[ $(uname -m) == 'x86_64' ]]; then
+        go build -o ../../../onionshare/resources/tor/meek-client ./meek-client
+    fi
 else
     go build -o ../../../onionshare/resources/tor/meek-client ./meek-client
 fi
